Replace clinit static puts with initial values for static fields
when a static put in clinit is dominated by both the exit block, and all
other usages of the static field, then static put can be removed, and the
dex file data section (static_values_off) holding the initial value for
static fields can be used for setting the initial value.

Include <clinit> as a callee in the call graph when a class is mentioned.

Also fix a bug in the reading for initial values for static fields
from dex files.

+++ src/main/java/com/android/tools/r8/utils/EncodedValueUtils.java
@@ -13,10 +13,11 @@ public class EncodedValueUtils {
     long result = 0;
     int shift = 0;
     for (int i = 1; i < numberOfBytes; i++) {
-      result |= ((long) ((file.get() & 0xFF))) << shift;
+      result |= ((long) (file.get() & 0xFF)) << shift;
       shift += 8;
     }
-    return result | (file.get() << shift);
+    // Let the last byte sign-extend into any remaining bytes.
+    return result | (((long) file.get()) << shift);
   }
 
   // Inspired by com.android.dex.EncodedValueCodec
